APPFACT-2087 : Modified the data type of the deviceminutes into Double from Long (#595)
* Updated the minutes type to double from long.
* Other fixes

+++ src/com/kony/appfactory/dto/tests/DetailedNativeResults.groovy
@@ -15,7 +15,7 @@ public class DetailedNativeResults implements Serializable {
       private String binaryExt;
       private long testDuration;
       private String resultsLink;
-      private long deviceMinutes;
+      private double deviceMinutes;
 
       public DetailedNativeResults() {
           this.device = new Device();
@@ -95,12 +95,12 @@ public class DetailedNativeResults implements Serializable {
       }
 
     @NonCPS
-    public long getDeviceMinutes() {
+    public double getDeviceMinutes() {
         return deviceMinutes
     }
 
     @NonCPS
-    public void setDeviceMinutes(long deviceMinutes) {
+    public void setDeviceMinutes(double deviceMinutes) {
         this.deviceMinutes = deviceMinutes
     }
